backend/requirements.txt
fastapi
uvicorn
transformers
torch
sqlalchemy
psycopg2-binary
python-dotenv
moviepy
tensorflow
huggingface-hub
passlib[bcrypt]
python-jose
python-multipart
pydantic
requests


